requirements.txt
 langchain_openai
 langchain_core
 python-dotenv
 streamlit
 langchain_community
 langserve
 langchain_ollama
 fastapi
 uvicorn
 sse_stralette
 bs4
 pypdf
 faiss-cpu
 wikipedia
 arxiv
 hub